Fix and display video
