+++ src/Camera/QGCCameraControl.cc
@@ -380,6 +380,8 @@ QGCCameraControl::takePhoto()
             _setPhotoStatus(PHOTO_CAPTURE_IN_PROGRESS);
             _captureInfoRetries = 0;
             //-- Capture local image as well
+#if 0
+            TODO Pass the hability to take  picture to Qml
             if(qgcApp()->toolbox()->videoManager()->videoReceiver()) {
                 QString photoPath = qgcApp()->toolbox()->settingsManager()->appSettings()->savePath()->rawValue().toString() + QStringLiteral("/Photo");
                 QDir().mkpath(photoPath);
@@ -387,6 +389,8 @@ QGCCameraControl::takePhoto()
                 qgcApp()->toolbox()->videoManager()->videoReceiver()->grabImage(photoPath);
             }
             return true;
+#endif
+
         }
     }
     return false;
@@ -1523,12 +1527,16 @@ QGCCameraControl::handleCaptureStatus(const mavlink_camera_capture_status_t& cap
     //-- Time Lapse
     if(photoStatus() == PHOTO_CAPTURE_INTERVAL_IDLE || photoStatus() == PHOTO_CAPTURE_INTERVAL_IN_PROGRESS) {
         //-- Capture local image as well
+
+#if 0
+        TODO: Pass the hability to Take a picture to Qml.
         if(qgcApp()->toolbox()->videoManager()->videoReceiver()) {
             QString photoPath = qgcApp()->toolbox()->settingsManager()->appSettings()->savePath()->rawValue().toString() + QStringLiteral("/Photo");
             QDir().mkpath(photoPath);
             photoPath += + "/" + QDateTime::currentDateTime().toString("yyyy-MM-dd_hh.mm.ss.zzz") + ".jpg";
             qgcApp()->toolbox()->videoManager()->videoReceiver()->grabImage(photoPath);
         }
+#endif
     }
 }
