+++ src/Vehicle/Vehicle.cc
@@ -1667,12 +1667,15 @@ void Vehicle::_updateArmed(bool armed)
             _trajectoryPoints->stop();
             _flightTimerStop();
             // Also handle Video Streaming
+#if 0
+            TODO: Emit a signal that should be connected to Qml
             if(qgcApp()->toolbox()->videoManager()->videoReceiver()) {
                 if(_settingsManager->videoSettings()->disableWhenDisarmed()->rawValue().toBool()) {
                     _settingsManager->videoSettings()->streamEnabled()->setRawValue(false);
                     qgcApp()->toolbox()->videoManager()->videoReceiver()->stop();
                 }
             }
+#endif
         }
     }
 }
