+++ src/VideoStreaming/VideoStreaming.cc
@@ -28,7 +28,7 @@
 #endif
 
 #include "VideoStreaming.h"
-#include "VideoController.h"
+#include "VideoReceiver.h"
 
 #if defined(QGC_GST_STREAMING)
     G_BEGIN_DECLS
@@ -169,6 +169,6 @@ void initializeVideoStreaming(int &argc, char* argv[], char* logpath, char* debu
     Q_UNUSED(debuglevel);
 #endif
     // Fake Player to start the video in the Rendering Thread.
-    qmlRegisterType<VideoController>              ("QGroundControl.QgcQtGStreamer", 1, 0, "VideoController");
+    qmlRegisterType<VideoReceiver>              ("QGroundControl.QgcQtGStreamer", 1, 0, "VideoReceiver");
 }
