+++ src/api/QGCCorePlugin.h
@@ -31,7 +31,6 @@ class QQmlApplicationEngine;
 class Vehicle;
 class LinkInterface;
 class QmlObjectListModel;
-class VideoReceiver;
 class PlanMasterController;
 class QGCCameraManager;
 class QGCCameraControl;
@@ -106,8 +105,8 @@ public:
 
     /// Allows the plugin to override the creation of VideoManager.
     virtual VideoManager* createVideoManager(QGCApplication* app, QGCToolbox* toolbox);
+
     /// Allows the plugin to override the creation of VideoReceiver.
-    virtual VideoReceiver* createVideoReceiver(QObject* parent);
 
     /// Allows the plugin to see all mavlink traffic to a vehicle
     /// @return true: Allow vehicle to continue processing, false: Vehicle should not process message
